+++ packages/core/src/components/document-component/document-component.tsx
@@ -65,6 +65,12 @@ export class DocumentComponent implements ComponentInterface {
   opschrift?: DocumentComponentInputType;
 
   /**
+   * The Kop XML.
+   */
+  @Prop()
+  kop?: DocumentComponentInputType;
+
+  /**
    * The Inhoud XML.
    */
   @Prop()
@@ -296,7 +302,21 @@ export class DocumentComponent implements ComponentInterface {
                 )}
                 <span id="heading-title">
                   {this.notApplicable && <span class="sr-only">Niet van toepassing:</span>}
-                  {this.label || this.nummer || this.opschrift ? (
+
+                  {this.kop && (
+                    <dso-ozon-content
+                      content={this.kop}
+                      addSpaceAfterNode
+                      onDsoAnchorClick={this.handleOzonContentAnchorClick}
+                      mark={this.mark && ((text) => this.mark?.(text, "kop"))}
+                      onDsoOzonContentMarkItemHighlight={(e) =>
+                        this.dsoMarkItemHighlight.emit({ ...e.detail, source: "kop" })
+                      }
+                      inline
+                    />
+                  )}
+
+                  {(this.label || this.nummer || this.opschrift) && !this.kop ? (
                     <>
                       {this.label && (
                         <>