+++ packages/core/src/components/ozon-content/nodes/kop.node.tsx
@@ -3,8 +3,8 @@ import { h, Fragment } from "@stencil/core";
 import { OzonContentNodeContext } from "../ozon-content-node-context.interface";
 import { OzonContentNode } from "../ozon-content-node.interface";
 
-export class OzonContentOpschriftNode implements OzonContentNode {
-  name = "Opschrift";
+export class OzonContentKopNode implements OzonContentNode {
+  name = ["Opschrift", "label", "nummer", "kop"];
 
   render(node: Element, { mapNodeToJsx }: OzonContentNodeContext) {
     let content = mapNodeToJsx(node.childNodes);